requirements.txt
hugchat
huggingface_hub
python-dotenv
langchain
PyQt5
